+++ prologin/semi_export.sh
@@ -1,11 +1,15 @@
 #!/bin/sh
 
 if [ $(id -u) -eq 0 ]; then
+  if [ "$#" -ne 1 ]; then
+    echo "please specify the file path to save the export to"
+    exit 1
+  fi
   echo -n "stopping website service..."
   systemctl stop website
   echo -e "\tDone !"
-  cd /home/prologin/
-  su prologin -c "pew in prologin-site sh $0 $1"
+  cd `dirname $0`
+  su prologin -c "pew in prologin-site sh `basename $0` `realpath $1`"
   systemctl start website
   exit 0
 fi
